refactor: projectDetail.tsxのonDragEndをhooksに切り分けた #182
