+++ frontend/src/pages/projectDetail/ProjectDetail.tsx
@@ -1,28 +1,17 @@
 import React, { FC, useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
-import { DropResult, resetServerContext } from 'react-beautiful-dnd'
+import { resetServerContext } from 'react-beautiful-dnd'
 import styled, { css } from 'styled-components'
-import { useAuthContext } from 'providers/AuthProvider'
-import { GetCurrentUserQuery } from './getUser.gen'
 import { useSearchSameCompanyUsersMutation } from '../projectList/projectList.gen'
-import {
-  useCreateInvitationMutation,
-  useUpdateTaskSortMutation,
-  useCreateListMutation,
-  useUpdateListSortMutation,
-} from './projectDetail.gen'
-import logger from 'utils/debugger/logger'
+import { useCreateInvitationMutation, useCreateListMutation } from './projectDetail.gen'
 import toast from 'utils/toast/toast'
 import { useInput } from 'hooks/useInput'
 import { useDebounce } from 'hooks/useDebounce'
 import { usePresence } from 'hooks/usePresence'
 import { useSetWeaponJson } from 'hooks/useSetWeaponJson'
 import { useCompleteAnimation } from 'hooks/useCompleteAnimation'
-import { StatusParam } from 'types/status'
 import { List } from 'types/list'
-import { Task } from 'types/task'
 import { DEFAULT_USER } from 'consts/defaultImages'
-import { DROP_TYPE } from 'consts/dropType'
 import { ProjectDrawer } from 'components/models/project/ProjectDrawer'
 import { ProjectRight } from 'components/models/project/ProjectRight'
 import { ProjectMyInfo } from 'components/models/project/ProjectMyInfo'
@@ -30,6 +19,7 @@ import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateSizeBasedOnFig
 import { ProjectDetailHeader } from 'components/ui/header/ProjectDetailHeader'
 import { LazyLoading } from 'components/ui/loading/LazyLoading'
 import { TaskCompleteAnimation } from 'components/models/task/animation/TaskCompleteAnimation'
+import { useProjectDetailDragEnd } from 'hooks/useProjectDetailDragEnd'
 import { useProjectDetail } from 'hooks/useProjectDetail'
 import { useGetCurrentUserData } from 'hooks/useGetCurrentUserData'
 import { Notifications } from 'types/notification'
@@ -40,7 +30,6 @@ export const ProjectDetail: FC = () => {
   resetServerContext()
   usePresence()
   const { id } = useParams()
-  const { currentUser } = useAuthContext()
   const { json, setWeapon } = useSetWeaponJson()
   const { anchorEl, isCompleted, setIsCompleted } = useCompleteAnimation<HTMLDivElement>(json)
   const [selectUserIds, setSelectUserIds] = useState<string[]>([])
@@ -52,9 +41,10 @@ export const ProjectDetail: FC = () => {
     setList,
   )
 
-  const { currentUserData, notifications } = useGetCurrentUserData()
+  const { currentUserData, firebaseCurrentUser, notifications } = useGetCurrentUserData()
 
   const [searchSameCompanyUsers, searchSameCompanyUsersData] = useSearchSameCompanyUsersMutation()
+
   const [createInvitation] = useCreateInvitationMutation({
     onCompleted(data) {
       toast.success(`${data.createInvitation.user.name}を招待しました`)
@@ -62,20 +52,6 @@ export const ProjectDetail: FC = () => {
     },
     onError(err) {
       toast.error('招待に失敗しました')
-    },
-  })
-
-  const [updateTaskSort] = useUpdateTaskSortMutation({
-    onCompleted(data) {
-      logger.table(data.updateTaskSort)
-      if (data.updateTaskSort.is_completed) {
-        setWeapon(data.updateTaskSort.high_status_name as StatusParam)
-        setIsCompleted(true)
-      }
-    },
-    onError(err) {
-      logger.debug(err)
-      toast.error('タスクの移動に失敗しました')
     },
   })
 
@@ -88,10 +64,12 @@ export const ProjectDetail: FC = () => {
     },
   })
 
-  const [updateList] = useUpdateListSortMutation({
-    onError(err) {
-      toast.error('リスト更新に失敗しました')
-    },
+  const { onDragEnd } = useProjectDetailDragEnd({
+    list,
+    setList,
+    setWeapon,
+    setIsCompleted,
+    firebaseCurrentUser,
   })
 
   const debouncedInputText = useDebounce<string>(inputUserName.value, 500)
@@ -110,124 +88,6 @@ export const ProjectDetail: FC = () => {
     setSelectUserIds(selectUserIds => [...selectUserIds, userId])
   }
 
-  const removeFromList = (list: Task[], index: number): [Task, Task[]] => {
-    const result = Array.from(list)
-    const [removed] = result.splice(index, 1)
-    return [removed, result]
-  }
-
-  const addToList = (list: Task[], index: number, element: Task) => {
-    const result = Array.from(list)
-    result.splice(index, 0, element)
-    return result
-  }
-
-  const reorder = (array: List[], startIndex: number, endIndex: number) => {
-    const result = Array.from(array)
-
-    const previousItem = result[endIndex]
-    const [removed] = result.splice(startIndex, 1)
-
-    if (previousItem) {
-      removed.index = previousItem.index
-    }
-    result.splice(endIndex, 0, removed)
-    return result
-  }
-
-  const onDragEnd = async (result: DropResult) => {
-    if (!result.destination) return
-    const { source, destination, type } = result
-    const destinationDroppableId = Number(destination.droppableId)
-    const sourceDroppableId = Number(source.droppableId)
-    const destinationIndex = destination.index
-    const sourceIndex = source.index
-    if (destinationDroppableId === sourceDroppableId && destinationIndex === sourceIndex) return
-
-    const listCopy = [...list]
-
-    if (type === DROP_TYPE.COLUMN) {
-      if (destinationIndex === 0) {
-        toast.warning('未着手は固定されています')
-        return
-      }
-      if (destinationIndex === list.length - 1) {
-        toast.warning('完了は固定されています')
-        return
-      }
-      const result = reorder(listCopy, sourceIndex, destinationIndex)
-      setList(result)
-      const updateListSort = result.map((list, index) => {
-        return {
-          id: list.sort_id,
-          task_list: index,
-        }
-      })
-      await updateList({
-        variables: {
-          listSort: updateListSort,
-        },
-      })
-      return
-    }
-
-    const [removedElement, newSourceList] = removeFromList(
-      listCopy[sourceDroppableId].tasks,
-      sourceIndex,
-    )
-
-    listCopy[sourceDroppableId].tasks = newSourceList
-
-    const destinationTask = listCopy[destinationDroppableId].tasks
-    listCopy[destinationDroppableId].tasks = addToList(
-      destinationTask,
-      destinationIndex,
-      removedElement,
-    )
-
-    const sortListCopy = listCopy.map(list => {
-      const sortList = list.tasks.map((task, index) => {
-        task.vertical_sort = index
-        return task
-      })
-
-      list.tasks = sortList
-      return list
-    })
-
-    //TODO: 完了にカードが移動したらcompleted_flgをtrueにする処理を書く必要あり
-    const updateTasks = sortListCopy.map((list, index, { length }) => {
-      logger.debug(length)
-      return list.tasks.map((task, taskIndex) => {
-        return {
-          id: task.id,
-          list_id: list.list_id,
-          vertical_sort: taskIndex,
-          completed_flg: index === length - 1 ? true : false,
-        }
-      })
-    })
-    // return
-
-    const joinUpdateTasks = []
-    for (let index = 0; index < updateTasks.length; index++) {
-      joinUpdateTasks.push(...updateTasks[index])
-    }
-
-    logger.table([...joinUpdateTasks])
-    await updateTaskSort({
-      variables: {
-        updateTasks: {
-          tasks: joinUpdateTasks,
-          project_id: String(id),
-          user_id: currentUser?.uid ?? '',
-        },
-      },
-    })
-
-    setList(listCopy)
-  }
-
   const handleCreateList = async () => {
     await createList({
       variables: {
@@ -235,7 +95,7 @@ export const ProjectDetail: FC = () => {
           name: 'リスト名',
           project_id: String(id),
           task_list: 1,
-          user_id: currentUser?.uid ?? '',
+          user_id: firebaseCurrentUser?.uid ?? '',
         },
       },
     })
